requirements.txt
PyMuPDF==1.24.0
python-docx==1.1.0
chromadb
google-genai
python-dotenv==1.0.1
playwright
beautifulsoup4
lxml
requests
anthropic
sentence-transformers
fastapi
uvicorn[standard]
sse-starlette
pydantic
